requirements.txt
python-dotenv==1.0.1
google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-binance==1.0.19
pydantic==2.8.2
pydantic-settings==2.4.0
python-telegram-bot==21.5
PyYAML==6.0.2
requests==2.31.0
fastapi==0.115.0
uvicorn==0.30.6
python-dateutil==2.9.0.post0
google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
schedule==1.2.2
pydantic==2.8.2
